requirements.txt
asgiref==3.3.4
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
fastapi==0.65.2
findspark==1.4.2
google-api-core==1.30.0
google-auth==1.31.0
google-cloud-pubsub==2.5.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.38.0
h11==0.12.0
httptools==0.2.0
idna==2.10
libcst==0.3.19
mypy-extensions==0.4.3
packaging==20.9
proto-plus==1.18.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==2.4.7
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.5
uvicorn==0.14.0
uvloop==0.15.2
watchgod==0.7
websockets==9.1
